+++ src/loop/branch.ts
@@ -1,0 +1,11 @@
+import type { BranchSample } from "../consistency.js";
+import type { BranchSummary } from "./types.js";
+
+export function summarizeBranch(chosen: BranchSample, samples: BranchSample[]): BranchSummary {
+  return {
+    budget: samples.length,
+    chosenIndex: chosen.index,
+    uncertainties: samples.map((s) => s.planState.uncertainties.length),
+    temperatures: samples.map((s) => s.temperature),
+  };
+}
